% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/canyon_wren-data.R
\docType{data}
\name{canyon_wren}
\alias{canyon_wren}
\title{Acoustic recording of a \emph{Catherpes mexicanus} (canyon wren) song.}
\format{One Wave object: \describe{
  
  \item{canyon_wren}{\emph{Catherpes mexicanus} recording}
  
  }}
\usage{
data(canyon_wren)
}
\description{
Acoustic recording of a \emph{Catherpes mexicanus} (canyon wren) song.
}
\keyword{datasets}
