% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynaSpec-package.R
\docType{package}
\name{dynaSpec}
\alias{dynaSpec}
\title{dynaSpec: Dynamic spectrogram visualization}
\description{
A set of tools to generate dynamic spectrogram visualizations in video format.
}
\details{
License: GPL (>= 2)
}
\author{
Marcelo Araya-Salas & Matthew Wilkins
  
  Maintainer: Marcelo Araya-Salas (\email{marceloa27@gmail.com})
}
